# Speed up `v2/pods/::status` (#6786)